W171d: Kosaka field family FOUR MATCHes + a window clear (101 -> 105 across three files). k_fldFrame 001AF930 (retail's '> 0xff' boundary form for the animation split); k_fldrc 001B56F0 (colour propagation rebuilt around RwRGBA with signed float-to-int channel construction inside the traversal plus base-first indexed access -- this also CLEARED its over-window) and 001B5E90 (vector colour scaling rebuilt around staged RwV3d data, a typed absolute symbol, separate created/existing target roles, and scaling applied AFTER all channel fetches); k_event 001C7E70 (event path format pointer typing/address materialization corrected, and the loader result actually returned on the load path).
Confirms the k_sceneDraw recipe transfers across the whole Kosaka field family: RwV3d/RwRGBA as real aggregates, packed colour construction ordered before scaling, base-first indexed access, separate roles rather than one reused variable. Zero lost, zero additions, markers unchanged in all three files, pragma brackets unchanged.

+++ src/Event/mt_evtMessage.c
@@ -8,7 +8,7 @@ extern u32 DAT_006a1850;
 extern u8 DAT_006a0000[];
 extern u32 DAT_006a1870;
 extern f32 DAT_007cae58;
-extern u32 DAT_007caf30;
+extern f32 DAT_007caf30;
 extern u32 DAT_007cd490;
 extern u32 DAT_007cd494;
 extern u32 DAT_007cd498;
